refactor: rename helpers.ts to common.ts
