+++ src/events/pull-request.ts
@@ -19,7 +19,7 @@ import {
   doCommit,
   doFilterSvgsFromFiles,
   doOptimizeSvgs,
-} from "./helpers";
+} from "./common";
 
 
 function getHeadRef(): string {
